feat: mirror deps from container to local machine
I want to have all of the project's deps managed in the container.

That means I do not want to install them locally.  But my IDE requires
that certain packages are installed locally (TS). To have the best of
both worlds, we can use a bind mount to mirror the node_modules in our
container back onto our local machine.