feat: keep node_modules dir but leave it empty
During development, I want all of the project's dependencies to be
installed in the container, not on my local machine.  However, I also
want to be able to make use of the dependencies on my local machine (for
example, my IDE requires that TS be installed/available locally).  To
achieve these two things, we use a bind mount to mirror our container's
node_modules directory onto our local host.  To enable this, we need an
existing node_modules directory on our local host.  Thus, we want to
keep the directory so the bind mount works when running docker compose
up, but we do not want to keep any of the node modules that are inside